Finished the transition
