+++ SpatialGDK/Source/SpatialGDKEditorToolbar/SpatialGDKEditorToolbar.Build.cs
@@ -27,6 +27,7 @@ public class SpatialGDKEditorToolbar : ModuleRules
                 "MessageLog",
                 "SpatialGDK",
                 "SpatialGDKEditor",
+                "SpatialGDKServices",
                 "UnrealEd"
             }
         );
